Added cooldowns and started splitting to SunnyUtils

+++ Plugin/src/main/java/me/partlysunny/command/annotations/Cooldown.kt
@@ -9,7 +9,7 @@ annotation class Cooldown(
     /**
      * Length of cooldown in milliseconds
      */
-    val length: Int,
+    val length: Long,
     /**
      * Message to display if cooldown is not over
      */
